Make batch number primary key of batch table (#198)
* Update models

* Update storage methods

* Update migrations files

* Make tests work

* Remove skip

* Uncomment example config

* Use batch ID in MarkBatchAsSubmitted

* Remove index on transaction hash field

* Add UNIQUE constraint on transaction_hash field

* Format migration file

* Refactor

+++ db/postgres/migrations/000002_commitment.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE commitment (
-    commitment_id      SERIAL PRIMARY KEY,
-    type               SMALLINT NOT NULL,
-    transactions       BYTEA    NOT NULL, -- 32 transactions encoded in compact format
-    fee_receiver       BIGINT   NOT NULL, -- state id of fee receiver
-    combined_signature BYTEA    NOT NULL,
-    post_state_root    BYTEA    NOT NULL,
-    included_in_batch  INT REFERENCES batch
+    commitment_id      SERIAL      PRIMARY KEY,
+    type               SMALLINT    NOT NULL,
+    transactions       BYTEA       NOT NULL, -- 32 transactions encoded in compact format
+    fee_receiver       BIGINT      NOT NULL, -- state id of fee receiver
+    combined_signature BYTEA       NOT NULL,
+    post_state_root    BYTEA       NOT NULL,
+    included_in_batch  NUMERIC(78) REFERENCES batch
 );
